# tests/api/openapi_v1_snapshot.json
{
  "components": {
    "schemas": {
...
      "CreateRoutineRequest": {
        "description": "The POST /v1/routines body: a workspace-scoped routine bound to a stored agent.\n\n``agent_id`` references a stored ``/v1/agents`` record in the same workspace. There is\ndeliberately NO ``agent_path`` field \u2014 a tenant may never name a server filesystem\npath (FS-leak surface). ``idempotency_key`` makes a re-POST return the prior routine.",
        "properties": {
          "agent_id": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Agent Id",
            "type": "string"
          },
          "deliver": {
            "default": "none",
            "enum": [
              "none",
              "telegram",
              "email"
            ],
            "title": "Deliver",
            "type": "string"
          },
          "enabled": {
            "default": true,
            "title": "Enabled",
            "type": "boolean"
          },
          "idempotency_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Idempotency Key"
          },
          "model": {
            "anyOf": [
              {
                "maxLength": 200,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          },
          "name": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "prompt": {
            "maxLength": 100000,
            "minLength": 1,
            "title": "Prompt",
            "type": "string"
          },
          "provider": {
            "anyOf": [
              {
                "maxLength": 100,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Provider"
          },
          "schedule": {
            "$ref": "#/components/schemas/Schedule"
          },
          "workspace_id": {
            "title": "Workspace Id",
            "type": "string"
          }
        },
        "required": [
          "workspace_id",
          "agent_id",
          "name",
          "prompt",
          "schedule"
        ],
        "title": "CreateRoutineRequest",
        "type": "object"
      },
...
      "ErrorResponse": {
        "description": "The consistent error envelope returned for handled failures (AAEO-9).",
        "properties": {
          "code": {
            "default": "error",
            "title": "Code",
            "type": "string"
          },
          "detail": {
            "title": "Detail",
            "type": "string"
          }
        },
        "required": [
          "detail"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "x-framework-generated": "validation-error-envelope"
      },
...
      "RoutineListResponse": {
        "description": "The GET /v1/routines envelope: the workspace's routines.",
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/himmy__api__routers__routines__RoutineView"
            },
            "title": "Items",
            "type": "array"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total"
        ],
        "title": "RoutineListResponse",
        "type": "object"
      },
...
      "himmy__api__routers__routines__RoutineView": {
        "description": "The API view of a stored routine (no ``agent_path`` \u2014 that field is local-only).",
        "properties": {
          "agent_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Agent Id"
          },
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "deliver": {
            "enum": [
              "none",
              "telegram",
              "email"
            ],
            "title": "Deliver",
            "type": "string"
          },
          "enabled": {
            "title": "Enabled",
            "type": "boolean"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "last_delivery": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Delivery"
          },
          "last_error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Error"
          },
          "last_preview": {
            "title": "Last Preview",
            "type": "string"
          },
          "last_run_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Run At"
          },
          "last_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Status"
          },
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "prompt": {
            "title": "Prompt",
            "type": "string"
          },
          "provider": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Provider"
          },
          "running": {
            "title": "Running",
            "type": "boolean"
          },
          "schedule": {
            "$ref": "#/components/schemas/Schedule"
          },
          "updated_at": {
            "title": "Updated At",
            "type": "string"
          },
          "workspace_id": {
            "title": "Workspace Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "workspace_id",
          "agent_id",
          "name",
          "prompt",
          "schedule",
          "provider",
          "model",
          "deliver",
          "enabled",
          "created_at",
          "updated_at",
          "last_run_at",
          "last_status",
          "last_preview",
          "last_error",
          "last_delivery",
          "running"
        ],
        "title": "RoutineView",
        "type": "object"
      },
...
    }
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/routines": {
      "get": {
        "description": "List the routines in a workspace (tenant-scoped), newest first.",
        "operationId": "list_routines_v1_routines_get",
        "parameters": [
          {
            "in": "query",
            "name": "workspace_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Workspace Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoutineListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Routines",
        "tags": [
          "routines"
        ]
      },
      "post": {
        "description": "Create a workspace-scoped routine bound to a stored agent (idempotent).\n\nThe ``agent_id`` must resolve to a stored agent in the same workspace (404 otherwise).\nA re-POST of an existing ``idempotency_key`` returns the prior routine unchanged.",
        "operationId": "create_routine_v1_routines_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRoutineRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/himmy__api__routers__routines__RoutineView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Routine",
        "tags": [
          "routines"
        ]
      }
    },
    "/v1/routines/{routine_id}": {
      "delete": {
        "description": "Delete a routine (tenant-scoped). 404 when absent / out-of-workspace.",
        "operationId": "delete_routine_v1_routines__routine_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "routine_id",
            "required": true,
            "schema": {
              "title": "Routine Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "workspace_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Workspace Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Delete Routine V1 Routines  Routine Id  Delete",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Resource not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Routine",
        "tags": [
          "routines"
        ]
      },
      "get": {
        "description": "Read one routine by id (404 when unknown / out-of-workspace).",
        "operationId": "get_routine_v1_routines__routine_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "routine_id",
            "required": true,
            "schema": {
              "title": "Routine Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "workspace_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Workspace Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/himmy__api__routers__routines__RoutineView"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Resource not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Routine",
        "tags": [
          "routines"
        ]
      },
      "patch": {
        "description": "Partial update of a routine (tenant-scoped). 404 when absent.\n\nA re-pointed ``agent_id`` is re-validated to exist in the workspace; the schedule, when\ngiven, is re-validated as a whole. ``agent_path`` cannot be set (no field).",
        "operationId": "update_routine_v1_routines__routine_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "routine_id",
            "required": true,
            "schema": {
              "title": "Routine Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRoutineRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/himmy__api__routers__routines__RoutineView"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Resource not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Routine",
        "tags": [
          "routines"
        ]
      }
    },
    "/v1/routines/{routine_id}/run-now": {
      "post": {
        "description": "Run a routine immediately through the unattended rails (tenant-scoped).\n\nDispatches via ``RunAppService.create_run`` under the routine's workspace (canonical\nstore + T0.4 quota + HITL pause). A routine already executing \u2014 in THIS process or in\nthe Studio scheduler process (the cross-process flock) \u2014 is refused with 409, never\nrun twice. The response carries the refreshed routine once the run settles/pauses.",
        "operationId": "run_now_v1_routines__routine_id__run_now_post",
        "parameters": [
          {
            "in": "path",
            "name": "routine_id",
            "required": true,
            "schema": {
              "title": "Routine Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "workspace_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Workspace Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/himmy__api__routers__routines__RoutineView"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Resource not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Run Now",
        "tags": [
          "routines"
        ]
      }
    },
...
  }
}